platform_code/requirements.txt
#Version of the packages used
Django==3.2.7
bleach===4.1.0
psycopg2-binary==2.8.4
gunicorn==20.0.4
django-bootstrap4==2.3.0
django-countries==7.2.1
django-markdownify==0.9.0
django-modeltranslation==0.16.1
django-ckeditor==6.1.0
requests==2.24.0
sentry-sdk==1.5.12
pyjwt==1.7.1
plotly==4.13.0
pillow==8.1.0
reportlab==3.5.59
pre-commit==2.20.0
